% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irwtls.R
\name{irwtls}
\alias{irwtls}
\title{Fitting iteratively reweighted least trimmed squares models}
\usage{
irwtls(formula, data, varVariable, dataset_identifier, iterations = 200,
  delta = 10^-4, tol = 10^-3)
}
\description{
Fitting iteratively reweighted least trimmed squares models
}
