% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emCode.R
\name{hhlm}
\alias{hhlm}
\title{Linear Regression for Data with Heteroskedastic Heavy-Tailed Data}
\usage{
hhlm(formula, varFormula, data, iterations = 20, delta = 10^-4)
}
\description{
Linear Regression for Data with Heteroskedastic Heavy-Tailed Data
}
